+++ gridcapa-swe-runner-api/pom.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xmlns="http://maven.apache.org/POM/4.0.0"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
         <artifactId>gridcapa-swe</artifactId>
         <groupId>com.farao-community.farao</groupId>
-        <version>0.1</version>
+        <version>0.0.1</version>
     </parent>
 
     <artifactId>gridcapa-swe-runner-api</artifactId>
     <name>GridCapa SWE runner API</name>
-    <version>0.1</version>
+    <version>0.0.1</version>
     <description>Client API to gridcapa swe process</description>
     <packaging>jar</packaging>
