+++ gridcapa-swe-runner-api/src/main/java/com/farao_community/farao/swe/api/resource/ThreadLauncherResult.java
@@ -4,11 +4,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-
 package com.farao_community.farao.swe.api.resource;
 
 import java.util.Optional;
 
+/**
+ * @author Theo Pascoli {@literal <theo.pascoli at rte-france.com>}
+ */
 public class ThreadLauncherResult<U> {
 
     private final Optional<U> result;
