+++ gridcapa-swe-runner-api/src/test/resources/sweRequestMessage.json
@@ -2,5 +2,5 @@
   "data": {
     "type": "swe-request",
     "id": "id"
-    }
+  }
 }
