--- gridcapa-swe-runner-app/src/main/java/com/farao_community/farao/swe/runner/app/configuration/EventsLogger.java
@@ -1,25 +1,0 @@
-/*
- * Copyright (c) 2022, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.farao_community.farao.swe.runner.app.configuration;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * @author Theo Pascoli {@literal <theo.pascoli at rte-france.com>}
- */
-@Configuration
-class EventsLogger {
-
-    @Bean
-    public Logger getLogger() {
-        return LoggerFactory.getLogger("BUSINESS_LOGGER");
-    }
-
-}
