+++ gridcapa-swe-runner-app/src/main/java/com/farao_community/farao/swe/runner/app/configurations/RequestConfiguration.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2022, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.farao_community.farao.swe.runner.app.configurations;
+
+import com.farao_community.farao.swe.runner.app.services.RequestService;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import reactor.core.publisher.Flux;
+
+import java.util.function.Function;
+
+/**
+ * @author Theo Pascoli {@literal <theo.pascoli at rte-france.com>}
+ */
+@Configuration
+public class RequestConfiguration {
+
+    @Bean
+    public Function<Flux<byte[]>, Flux<byte[]>> request(RequestService requestService) {
+        return sweRequestFlux -> sweRequestFlux
+                .map(requestService::launchSweRequest)
+                .log();
+    }
+
+}
