+++ gridcapa-swe-runner-app/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,0 +1,9 @@
+{
+  "properties": [
+    {
+      "name": "swe-runner.bindings.request.destination",
+      "type": "java.lang.String",
+      "description": "Description for swe-runner.bindings.request.destination."
+    }
+  ]
+}
