+++ gridcapa-swe-runner-app/src/test/java/com/farao_community/farao/swe/runner/app/SweApplicationTest.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) 2022, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.farao_community.farao.swe.runner.app;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+/**
+ * @author Theo Pascoli {@literal <theo.pascoli at rte-france.com>}
+ */
+@SpringBootTest
+public class SweApplicationTest {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
