+++ gridcapa-swe-runner-starter/src/main/java/com/farao_community/farao/swe/runner/starter/SweClientProperties.java
