+++ gridcapa-swe-runner-starter/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,0 +1,24 @@
+{
+  "properties": [
+    {
+      "name": "swe-runner.binding.application-id",
+      "type": "java.lang.String",
+      "description": "Name of the application that call SWE through starter."
+    },
+    {
+      "name": "swe-runner.binding.destination",
+      "type": "java.lang.String",
+      "description": "The exchange name on which to send the SWE request."
+    },
+    {
+      "name": "swe-runner.binding.routing-key",
+      "type": "java.lang.String",
+      "description": "The routing key to use for the binding to swe-runner instance. Default value \"#\" is used to notify all queues."
+    },
+    {
+      "name": "swe-runner.binding.expiration",
+      "type": "java.lang.String",
+      "description": "Expiration time of SWE requests."
+    }
+  ]
+}
