+++ gridcapa-swe-runner-starter/src/main/resources/META-INF/spring.factories
