+++ gridcapa-swe-runner-starter/src/test/resources/sweRequestMessage.json
