+++ gridcapa-swe-runner-starter/src/test/resources/sweResponseMessage.json
